src/main/webapp/app/entities/concierge/route/concierge-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IConcierge, Concierge } from '../concierge.model';
import { ConciergeService } from '../service/concierge.service';

import { ConciergeRoutingResolveService } from './concierge-routing-resolve.service';

describe('Concierge routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: ConciergeRoutingResolveService;
    let service: ConciergeService;
    let resultConcierge: IConcierge | undefined;

    beforeEach(() => {
        TestBed.configureTestingModule({
            imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
            providers: [
                {
                    provide: ActivatedRoute,
                    useValue: {
                        snapshot: {
                            paramMap: convertToParamMap({}),
                        },
                    },
                },
            ],
        });
        mockRouter = TestBed.inject(Router);
        jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
        mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
        routingResolveService = TestBed.inject(ConciergeRoutingResolveService);
        service = TestBed.inject(ConciergeService);
        resultConcierge = undefined;
    });

    describe('resolve', () => {
        it('should return IConcierge returned by find', () => {
            // GIVEN
            service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
            mockActivatedRouteSnapshot.params = { id: 123 };

            // WHEN
            routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
                resultConcierge = result;
            });

            // THEN
            expect(service.find).toBeCalledWith(123);
            expect(resultConcierge).toEqual({ id: 123 });
        });

        it('should return new IConcierge if id is not provided', () => {
            // GIVEN
            service.find = jest.fn();
            mockActivatedRouteSnapshot.params = {};

            // WHEN
            routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
                resultConcierge = result;
            });

            // THEN
            expect(service.find).not.toBeCalled();
            expect(resultConcierge).toEqual(new Concierge());
        });

        it('should route to 404 page if data not found in server', () => {
            // GIVEN
            jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Concierge })));
            mockActivatedRouteSnapshot.params = { id: 123 };

            // WHEN
            routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
                resultConcierge = result;
            });

            // THEN
            expect(service.find).toBeCalledWith(123);
            expect(resultConcierge).toEqual(undefined);
            expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
        });
    });
});
